feat: all changed entired :sparkles:

+++ app/src/main/java/com/furkanhrmnc/filmscape/domain/model/Movie.kt
@@ -15,6 +15,6 @@ data class Movie(
     val posterPath: String,
     val releaseDate: String,
     val title: String,
-    val rating: Double,
-    val vote: Int,
+    val rating: Long,
+    val vote: Float,
 ) : Parcelable
